Overhaul on entity structure and resources class. Added clothing and factions.

+++ src/Time.cc
@@ -1,7 +1,7 @@
 #include "Time.h"
 
 sf::Clock Time::timer{};
-double Time::deltaTime{0};
+float Time::deltaTime{0};
 
 Time::Time()
 {
